requirements.txt
mysql_connector_repackaged==0.3.1
python-dotenv==0.19.2
requests==2.25.1
selenium==4.1.0
